+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - REMIX-3657: Added a Categories Interaction Tab to the Stage Manager
 - REMIX-3767: Added a Meshes Tab for the Stage Manager
 - REMIX-3659: Add the ability to get default output directory in Ingestion and AI Tools tabs
+- REMIX-3770: Added a Materials Tab for the Stage Manager
 
 ### Changed
 - Updated hdremix to dd92d0f